[COZY-46] refactor: return 코드 수정, 테스트 클래스 위치 수정
